requirements.txt
flask
gunicorn
openai
psycopg2-binary
pgvector
python-dotenv
beautifulsoup4
lxml
langchain-text-splitters
